pyproject.toml
[tool.poetry]
name = "money-tracker-server"
version = "0.1.0"
description = ""
authors = ["Your Name <you@example.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.13"
django = "5.1.7"
djangorestframework = "^3.15.2"
django-cors-headers = "^4.7.0"
dj-rest-auth = "^7.0.1"
python-dotenv = "^1.0.1"
psycopg2-binary = "^2.9.10"
pandas = "^2.2.3"
djangorestframework-simplejwt = "^5.5.0"
drf-yasg = "^1.21.10"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
